move start_blocking_server to SDK instead
