+++ oak_channel/Cargo.toml
@@ -7,15 +7,13 @@ license = "Apache-2.0"
 
 [features]
 default = []
-std = ["oak_channel_core/std"]
+std = []
 client = ["std"]
 
 [dependencies]
 anyhow = { version = "*", default-features = false }
 log = { version = "*", default-features = false }
-micro_rpc = { workspace = true }
 oak_core = { workspace = true }
-oak_channel_core = { workspace = true }
 static_assertions = "*"
 bitflags = "*"
 bytes = { version = "*", default-features = false }
