+++ oak_channel/src/lib.rs
@@ -15,17 +15,16 @@
 //
 
 #![cfg_attr(not(feature = "std"), no_std)]
-#![feature(never_type)]
-#![feature(unwrap_infallible)]
 #![allow(rustdoc::private_intra_doc_links)]
+#![feature(array_chunks)]
 
 #[cfg(feature = "client")]
 pub mod client;
 
-pub mod server;
-
+pub mod basic_framed;
 mod frame;
-mod message;
+pub mod message;
+pub mod server;
 
 #[cfg(test)]
 mod tests;
@@ -35,8 +34,44 @@ use alloc::{boxed::Box, vec::Vec};
 
 use anyhow::Context;
 use bytes::BytesMut;
-pub use oak_channel_core::{Channel, Read, Write};
 use oak_core::timer::Timer;
+
+/// Simple no_std compatible equivalent of [`std::io::Read`].
+///
+/// [`std::io::Read`]: <https://doc.rust-lang.org/std/io/trait.Read.html>
+pub trait Read {
+    fn read_exact(&mut self, data: &mut [u8]) -> anyhow::Result<()>;
+}
+
+#[cfg(feature = "std")]
+impl<T: std::io::Read> Read for T {
+    fn read_exact(&mut self, data: &mut [u8]) -> anyhow::Result<()> {
+        self.read_exact(data).map_err(anyhow::Error::msg)
+    }
+}
+
+/// Simple no_std compatible equivalent of [`std::io::Write`].
+///
+/// [`std::io::Write`]: <https://doc.rust-lang.org/std/io/trait.Write.html>
+pub trait Write {
+    fn write_all(&mut self, data: &[u8]) -> anyhow::Result<()>;
+    fn flush(&mut self) -> anyhow::Result<()>;
+}
+
+#[cfg(feature = "std")]
+impl<T: std::io::Write> Write for T {
+    fn write_all(&mut self, data: &[u8]) -> anyhow::Result<()> {
+        self.write_all(data).map_err(anyhow::Error::msg)
+    }
+
+    fn flush(&mut self) -> anyhow::Result<()> {
+        std::io::Write::flush(self).map_err(anyhow::Error::msg)
+    }
+}
+
+pub trait Channel: Read + Write + Send + Sync {}
+
+impl<T: Read + Write + Send + Sync> Channel for T {}
 
 struct InvocationChannel {
     inner: frame::Framed,
