+++ oak_restricted_kernel_sdk/Cargo.toml
@@ -23,5 +23,6 @@ oak_restricted_kernel_dice = { workspace = true, optional = true }
 oak_stage0_dice = { workspace = true, optional = true }
 p256 = { version = "*", default-features = false, features = ["ecdsa"] }
 strum = { version = "*", default-features = false, features = ["derive"] }
+micro_rpc = { workspace = true }
 lazy_static = { version = "*", features = ["spin_no_std"] }
 zerocopy = "*"
