+++ oak_restricted_kernel_sdk/src/lib.rs
@@ -15,6 +15,8 @@
 //
 
 #![cfg_attr(not(test), no_std)]
+#![feature(never_type)]
+#![feature(unwrap_infallible)]
 
 extern crate alloc;
